Corrected heresiesInHistoricalAreas (#740)
Moved the links for priscillianism and valentinianism to the top of # The Christian Schism section, and removed the heresiesInHistoricalAreas=yes switch added to the nestorian link.

We should leave the conversion to Nestorianism in the religion_map as the default for eastern areas that were historically under the Patriarch of the East, rather than sequestering that into the heresiesInHistoricalAreas=yes option. The Church of the East ratified the Council of Chalcedon in 554 under the PoE, Mar Aba I, even though it remained a distinct particular church. https://en.wikipedia.org/wiki/Aba_I It is not a small heretical sect and was of major historical reach and influence even into the eras of rule under the Islamic caliphates.

I posted a suggestion in the forums a while back about how it's somewhat misrepresented in the game, but it would be more ahistorical to make those areas Orthodox in earlier centuries, and categorize the Church of the East as heretical vis-a-vis the Catholics and Orthodox, but not the Copts.  https://forum.paradoxplaza.com/forum/threads/please-give-ecumenism-to-nestorianism-and-possibly-change-name-to-syriac.1475831/