Various Edits (Cultures, WRE Title, TFE law fix, etc.) (#2573) #minor
Closes #2373, closes #2400
- Added `e_western_roman_empire` title, and gave it proper title
mappings.

- Modified various decisions/effects/etc to account for new
`e_western_roman_empire` title. (There is likely still a few things that
could/should be modified, but I think I got most of the important ones
for now)

- Added custom COA definitions for `e_western_roman_empire` and
`e_byzantium` based off the Imperator Rome flags.

- Modified culture mapping for CK3 `roman` so that any Latin/Italic
cultures that are in Imperator's ROM, ERE, or WRE will now become Roman,
instead of just under ROM.

Closes #2112
- Add new non-Romanized Greeks culture, `hellenic_greek`, and made that
the fallback culture mapping for Imperator greek cultures.

- Made some of the converter cultures descend from `hellenic_greek`
instead of `greek`.

- Made a new definition of `greek` in converter_cultures.txt so it will
overwrite the base game definition to properly show that it is now a
hybrid of `roman` and `hellenic_greek`.

- Renamed `greek` to Romaios.

- Made it so CK3's base game `greek` culture only gets mapped to by
Greek cultures that are in ROM, ERE, or WRE in Imperator.

Closes #2500
- Modified the can_pass conditions for TFE's imperial_authority_3 law to
require that at least one year has passed in-game. Normally TFE checks
each of its start dates separately to see if at least 5 days has passed,
but that isn't compatible with the converter's dynamic start date, so it
had to be changed, and this seems to be the best way to make it
compatible with any start date.

- Added localization to explain the delay in game so users aren't
confused why they can't pass the law.


Other:
- Updated `roman` culture definition to use RtP traditions, where
applicable.

- Modified cultures that use `heritage_byzantine` to instead use TFE's
`heritage_hellenistic` if TFE is loaded so they are all in the same
heritage.

- Noticed a few cultures that had the same colors set, making them blend
together. Changed a few to better differentiate them.

- Added a game start on_action that replaces the primary title of any
titular duchy-level admin governor to use a relevant duchy that actually
has counties beneath it to ensure the admin mechanics work properly.

- Changed Imperator's `syriote` culture to map to `syro_hellenic` to be
more in line with TFE's mappings.

- Made `aremorican` map alongside rest of the gallic cultures.

- Made the Britannic cultures map to `breton` if they are in the
armorica_region.

---------

Co-authored-by: iht <IhateTrains@users.noreply.github.com>