Conversion of Imperator Defensive Leagues + compatches for TFE and RoA (#2631)
### CHANGES
- Finished game_start effect that sets up converted Imperator Defensive
Leagues. (Closes #2591)
- I decided to not add the Confederation/Defensive League split stuff
directly into the converter because it ended up requiring quite a bit of
changes that could be annoying to update whenever something needs to be
added or changed. Because it is something that can be used without the
converter too, I kept it as a separate mod that I uploaded to the Steam
Workshop called "Confederations & Leagues".
- I modified the welcome event that gets triggered at game start for
players (that normally notifies them of the game rules added by the
converter), and added a section to explain any/all mods that end up
being designed to work alongside the converter, so players can know that
the mod exists, and what it does.
- I setup the game_start effect so that if it detects that the
"Confederations & Leagues" mod is loaded (through a global variable), it
will properly setup the converted Defensive Leagues according to how the
mod allows them to be setup. If it detects that you are playing without
the mod, then it will only setup the converted Defensive Leagues that
can properly be setup in base game CK3 (meaning only Tribes/Nomads below
kingdom-tier can get in).

- Made a global variable (`irtock3_enabled`) get set at game start that
indicates that the converter was used (and you are currently using a
converted save game), so that if any mod in the future wants/needs to
consider whether you are playing a converted Imperator game, they can
easily check that variable.

- Updated the files for the new TFE update (didn't add anything new
aside from just making sure there were no replaceable/removable code
errors) (Closes #2639)


### UPDATE (June 20)
- Updated for RoA Version 4.0.0 (mainly just made sure there were no big
errors when converting, haven't added much of their new content yet)

- RoA changed how it handles the CCU parameter to be similar to how
WtWSMS handles it, so the relevant code files were modified to account
for that change.

- Converter heritage and language files were modified to account for RoA
CCU change, as well as change some of the parameters given to some of
the heritages/languages by the mods so they make a little more sense.

- Merged the Venetic heritage and language file into the main
IRToCK3_heritage/language files to simplify things.

- Made Basque have Equal Inheritance instead of Visigothic Codes in
vanilla CK3 and RoA (TFE and WtWSMS make their own edits)

- Renamed zzz_IRToCK3_tfe_override_l_english.yml.liquid to
zzz_IRToCK3_mod_override_l_english.yml.liquid and added a section for
ROA.

- Made Talaiotic map to Nuragic, since they were expected to be closer
to them than the Basque.

- Added a mapping to ROA's new Nenets culture in the proper out-of-scope
region of CK3

- Added a mapping for the romano_hellenic and helleno_roman cultures
from the Imperator "Terra Europea - Terra Indomita Sub-Mod" to greek
since they represent a kind of Roman-Hellenic hybrid culture in
Imperator

- Redid the province mappings for RoA since they changed their map a
bit. Might need to be reworked later, but should be fine for now. (Since
the converter can now convert wasteland provinces to whoever owns most
of the land surrounding them, I left quite a bit of the wasteland
province mappings the provinceMapper made)

- Redid the imperator_invictus mappings, removing the mapped provinces
from Subsaharan Africa so they don't all end up Berber/Megalithic.


### UPDATE (June 21)
- Moved 00_mongol_invasion_effects.txt edit from the TFE versions of
replaceable/removable_file_blocks to vanilla version of files.

- Modified become_chakravarti_decision and unite_india_decision_effect
so that they can only be taken if e_india has no holder or de jure land,
and instead of requiring the three normal Indian Empire titles, it
instead requires you to control the entire world_india region (pretty
much the same thing as what it required before), and it de jure drifts
the kingdoms of all empires that are 80% inside that india region.
(Closes #2596)

- Changed title mappings so that BHA maps to e_india instead of MRY.
e_india and BHA both represent a united India title/tag in their
respective games, so it makes more sense to have those mapped together.

- Made HIB map to k_ireland instead of IVE since HIB actually represents
a united Ireland.

- Updated some of the replaceable/removable_file_blocks for ROA and TFE.

---------

Co-authored-by: iht <IhateTrains@users.noreply.github.com>